finetune IPG with sim on-policy and hardware off-policy, rewards increase!

+++ scripts/eval_rewards.py
@@ -10,11 +10,11 @@ import pandas as pd
 import matplotlib.pyplot as plt
 
 ROOT_DIR = f"{LEGGED_GYM_ROOT_DIR}/logs/mini_cheetah_ref/"
-LOAD_RUN = "Jul13_01-49-59_PPO32_S16"
+LOAD_RUN = "Jul12_15-53-57_IPG32_S16"
 REWARDS_FILE = "rewards.csv"
-ITERATIONS = range(700, 705)
 
-PLOT = True
+DATAFRAME_LOADED = True
+PLOT = False
 PLOT_N = 1000  # number of steps to plot
 
 # Data struct fields from Robot-Software logs
@@ -36,14 +36,14 @@ DATA_LIST = [
 ]
 
 REWARD_WEIGHTS = {
+    "tracking_lin_vel": 4.0,
+    "tracking_ang_vel": 2.0,
     "min_base_height": 1.5,
-    # "tracking_lin_vel": 4.0,
-    # "tracking_ang_vel": 2.0,
     "orientation": 1.0,
-    "swing_grf": 1.0,
-    "stance_grf": 1.0,
-    "stand_still": 1.0,
-    "action_rate": 0.001,
+    "stand_still": 2.0,
+    "swing_grf": 3.0,
+    "stance_grf": 3.0,
+    "action_rate": 0.01,
     "action_rate2": 0.001,
 }
 
@@ -110,21 +110,78 @@ def get_rewards(it, data_struct, rewards_path):
 
 def setup(name="SMOOTH_RL_CONTROLLER"):
     data_dict = {}
-    for it in ITERATIONS:
-        path = os.path.join(ROOT_DIR, LOAD_RUN, f"{it}.mat")
+    log_files = [
+        file
+        for file in os.listdir(os.path.join(ROOT_DIR, LOAD_RUN))
+        if file.endswith(".mat")
+    ]
+    log_files = sorted(log_files)
+    for file in log_files:
+        iteration = int(file.split(".")[0])
+        path = os.path.join(ROOT_DIR, LOAD_RUN, file)
         data = scipy.io.loadmat(path)
-        data_dict[it] = data[name][0][0]
+        data_dict[iteration] = data[name][0][0]
     return data_dict
 
 
 if __name__ == "__main__":
     data_dict = setup()
     rewards_path = os.path.join(ROOT_DIR, LOAD_RUN, REWARDS_FILE)
-    if os.path.exists(rewards_path):
-        os.remove(rewards_path)
 
-    for it, data_struct in data_dict.items():
-        get_rewards(it, data_struct, rewards_path)
+    if not DATAFRAME_LOADED:
+        # Compute rewards and store in dataframe
+        if os.path.exists(rewards_path):
+            os.remove(rewards_path)
+        for it, data_struct in data_dict.items():
+            get_rewards(it, data_struct, rewards_path)
 
     rewards_df = pd.read_csv(rewards_path)
     print(rewards_df)
+
+    # Plot rewards stats
+    num_plots = len(REWARD_WEIGHTS) + 1  # +1 for total rewards
+    cols = 5
+    rows = np.ceil(num_plots / cols).astype(int)
+    fig, axs = plt.subplots(rows, cols, figsize=(15, 6))
+    fig.suptitle("IPG Finetuning Rewards")
+    axs = axs.flatten()
+
+    for i, key in enumerate(REWARD_WEIGHTS.keys()):
+        rewards_mean = rewards_df[rewards_df["type"] == key]["mean"].reset_index(
+            drop=True
+        )
+        rewards_std = rewards_df[rewards_df["type"] == key]["std"].reset_index(
+            drop=True
+        )
+        axs[i].plot(rewards_mean, label=key)
+        axs[i].fill_between(
+            range(len(rewards_mean)),
+            rewards_mean - rewards_std,
+            rewards_mean + rewards_std,
+            color="b",
+            alpha=0.2,
+        )
+        axs[i].legend()
+        axs[i].set_xlabel("Iter")
+
+    i = num_plots - 1
+    total_mean = rewards_df[rewards_df["type"] == "total"]["mean"].reset_index(
+        drop=True
+    )
+    total_std = rewards_df[rewards_df["type"] == "total"]["std"].reset_index(drop=True)
+    axs[i].plot(total_mean, label="total", color="r")
+    axs[i].fill_between(
+        range(len(total_mean)),
+        total_mean - total_std,
+        total_mean + total_std,
+        color="r",
+        alpha=0.2,
+    )
+    axs[i].legend()
+    axs[i].set_xlabel("Iter")
+
+    for i in range(num_plots, len(axs)):
+        axs[i].axis("off")
+
+    plt.tight_layout()
+    plt.show()
